requirements.txt
numpy==1.26.4
matplotlib==3.8.4
tqdm==4.66.4
scipy==1.13.1
pylatexenc==2.10
qiskit==1.2.2
qiskit-aer==0.14.2
jupyterlab==4.2.5
notebook==7.2.1
PyYAML==6.0.2

--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121
